MBS-11330: Split root/types.js into smaller files
