+++ root/search/types.js
@@ -32,3 +32,15 @@ export type ResultsPropsWithContextT<T> = {
   +form: SearchFormT,
   +lastUpdated?: string,
 };
+
+export type SearchResultT<T> = {
+  +entity: T,
+  +extra: $ReadOnlyArray<{
+    +medium_position: number,
+    +medium_track_count: number,
+    +release: ReleaseT,
+    +track_position: number,
+  }>,
+  +position: number,
+  +score: number,
+};
