+++ root/taglookup/types.js
@@ -7,6 +7,8 @@
  * later version: http://www.gnu.org/licenses/gpl-2.0.txt
  */
 
+import {type SearchResultT} from '../search/types';
+
 export type TagLookupFormT = FormT<{
   +artist: ReadOnlyFieldT<string>,
   +duration: ReadOnlyFieldT<string>,
