+++ root/types/autoeditorelection.js
@@ -1,0 +1,37 @@
+/*
+ * @flow strict
+ * Copyright (C) 2017 MetaBrainz Foundation
+ *
+ * This file is part of MusicBrainz, the open internet music database,
+ * and is licensed under the GPL version 2, or (at your option) any
+ * later version: http://www.gnu.org/licenses/gpl-2.0.txt
+ */
+
+// MusicBrainz::Server::Entity::AutoEditorElection::TO_JSON
+declare type AutoEditorElectionT = {
+  ...EntityRoleT<empty>,
+  +candidate: EditorT,
+  +close_time?: string,
+  +current_expiration_time: string,
+  +is_closed: boolean,
+  +is_open: boolean,
+  +is_pending: boolean,
+  +no_votes: number,
+  +open_time?: string,
+  +propose_time: string,
+  +proposer: EditorT,
+  +seconder_1?: EditorT,
+  +seconder_2?: EditorT,
+  +status_name: string,
+  +status_name_short: string,
+  +votes: $ReadOnlyArray<AutoEditorElectionVoteT>,
+  +yes_votes: number,
+};
+
+// MusicBrainz::Server::Entity::AutoEditorElectionVote::TO_JSON
+declare type AutoEditorElectionVoteT = {
+  ...EntityRoleT<empty>,
+  +vote_name: string,
+  +vote_time: string,
+  +voter: EditorT,
+};
